added admin.py file to respectve domains
- this reducing need to go to core/admin.py to modify admin files
- respective admin file likes near it's models
