+++ core/admin.py
@@ -20,63 +20,6 @@ from users.models.users import User
 # Simple models
 
 
-class ReturnRequestItemInline(admin.TabularInline):
-    model = ReturnRequestItem
-    extra = 0
-    readonly_fields = ("item_type", "quantity", "status")
-
-
-@admin.register(ReturnRequest)
-class ReturnRequestAdmin(admin.ModelAdmin):
-    list_display = (
-        "id",
-        "requester",
-        "status",
-        "requested_at",
-        "processed_at",
-    )
-
-    list_filter = (
-        "status",
-        "requested_at",
-    )
-
-    search_fields = (
-        "id",
-        "requester__email",
-        "requester__username",
-    )
-
-    readonly_fields = (
-        "requested_at",
-        "processed_at",
-    )
-
-    inlines = [ReturnRequestItemInline]
-
-
-@admin.register(ReturnRequestItem)
-class ReturnRequestItemAdmin(admin.ModelAdmin):
-    list_display = (
-        "id",
-        "return_request",
-        "item_type",
-        "quantity",
-        "status",
-    )
-
-    list_filter = (
-        "status",
-        "item_type",
-    )
-
-    search_fields = (
-        "return_request__id",
-    )
-
-    readonly_fields = (
-        "return_request",
-    )
 
 @admin.register(SecuritySettings)
 class SecuritySettingsAdmin(admin.ModelAdmin):
@@ -291,43 +234,12 @@ class ScheduledTaskRunAdmin(admin.ModelAdmin):
         "message",
     )
 
-admin.site.register(EquipmentAssignment)
-
-admin.site.register(EquipmentEvent)
 
 admin.site.register(SiteNameChangeHistory)
 
 admin.site.register(SiteRelocationHistory)
 
-admin.site.register(Department)
 
-@admin.register(UserPlacement)
-class UserPlacementAdmin(admin.ModelAdmin):
-    list_display = ("public_id", "user_full_name", "user_public_id", "room_public_id", "date_joined")
-    search_fields = (
-        "user__fname",
-        "user__lname",
-        "user__public_id",
-        "room__public_id",
-    )
-
-
-    def user_full_name(self, obj):
-        return f"{obj.user.fname} {obj.user.lname}"
-    user_full_name.short_description = "User Name"
-
-    def user_public_id(self, obj):
-        return obj.user.public_id
-    user_public_id.short_description = "User Public ID"
-
-    def room_public_id(self, obj):
-        return obj.room.public_id if obj.room else "-"
-    room_public_id.short_description = "Room Public ID"
-    
-@admin.register(Accessory)
-class AccessoryAdmin(admin.ModelAdmin):
-    list_display = ("public_id", "name", "quantity", "room", "is_deleted", "deleted_at")
-    search_fields = ("public_id", "name")
 
 
 @admin.register(UserSession)
@@ -338,186 +250,6 @@ class UserSessionAdmin(admin.ModelAdmin):
 
     ordering = ("-created_at",)  # "-" means descending order
 
-@admin.register(RoleAssignment)
-class RoleAssignmentAdmin(admin.ModelAdmin):
-    list_display = ('user', 'role', 'department', 'location', 'room', 'assigned_by', 'assigned_date', 'public_id')
-    readonly_fields = ('public_id',)
-    search_fields = ('user__email', 'role', 'department__name', 'location__name', 'room__name', 'assigned_by__email')
-    list_filter = ('role', 'department')
-
-@admin.register(User)
-class UserAdmin(BaseUserAdmin):
-
-    ordering = ("email",)
-
-    list_display = (
-        "email",
-        "fname",
-        "lname",
-        "active_role",
-        "public_id",
-        "is_active",
-        "is_locked",
-        "failed_login_attempts",
-        "locked_until",
-        "created_by",
-        "is_system_user",
-        "force_password_change",
-    )
-
-    list_filter = (
-        "is_active",
-        "is_locked",
-        "is_staff",
-        "is_superuser",
-        "is_system_user",
-        "force_password_change",
-    )
-
-    search_fields = (
-        "email",
-        "fname",
-        "lname",
-        "public_id",
-    )
-
-    readonly_fields = (
-        "public_id",
-        "failed_login_attempts",
-        "last_failed_login_at",
-        "locked_until",
-    )
-
-    fieldsets = (
-        (
-            None,
-            {
-                "fields": (
-                    "email",
-                    "password",
-                )
-            },
-        ),
-
-        (
-            "Personal info",
-            {
-                "fields": (
-                    "fname",
-                    "lname",
-                    "job_title",
-                    "role",
-                )
-            },
-        ),
-
-        (
-            "Permissions",
-            {
-                "fields": (
-                    "is_active",
-                    "is_staff",
-                    "is_superuser",
-                    "groups",
-                    "user_permissions",
-                )
-            },
-        ),
-
-        (
-            "Security",
-            {
-                "fields": (
-                    "is_locked",
-                    "locked_reason",
-                    "failed_login_attempts",
-                    "last_failed_login_at",
-                    "locked_until",
-                    "force_password_change",
-                )
-            },
-        ),
-
-        (
-            "System",
-            {
-                "fields": (
-                    "public_id",
-                    "created_by",
-                    "is_system_user",
-                )
-            },
-        ),
-
-        (
-            "Important dates",
-            {
-                "fields": (
-                    "last_login",
-                    "date_joined",
-                )
-            },
-        ),
-    )
-
-    add_fieldsets = (
-        (
-            None,
-            {
-                "classes": ("wide",),
-                "fields": (
-                    "email",
-                    "fname",
-                    "lname",
-                    "password1",
-                    "password2",
-                    "is_staff",
-                    "is_superuser",
-                ),
-            },
-        ),
-    )
-
-@admin.register(Room)
-class RoomAdmin(admin.ModelAdmin):
-    list_display = ('name', 'location')
-    readonly_fields = ("public_id",)
-    search_fields = ('name', 'location__name', 'public_id')
-
-
-@admin.register(Equipment)
-class EquipmentAdmin(admin.ModelAdmin):
-    list_display = ('name', 'brand', 'serial_number', 'public_id')  
-    readonly_fields = ('public_id',) 
-    search_fields = (
-        "public_id",
-        "name",
-        "brand",
-        "model",
-    )
-    fields = ('name', 'brand', 'serial_number', 'model', 'public_id', 'room', 'status', 'is_deleted', 'deleted_at')
-
-
-@admin.register(Component)
-class ComponentAdmin(admin.ModelAdmin):
-    list_display = ('name', 'brand', 'model', 'public_id')
-    readonly_fields = ('public_id',)
-    fields = ('name', 'brand', 'quantity', 'serial_number', 'model', 'public_id', 'equipment')
-
-
-@admin.register(Location)
-class LocationAdmin(admin.ModelAdmin):
-    list_display = ('name', 'department', 'public_id')  # adjust based on your model fields
-    readonly_fields = ('public_id',)
-    fields = ('name', 'department', 'public_id') 
-
-
-@admin.register(Consumable)
-class ConsumableAdmin(admin.ModelAdmin):
-    list_display = ('name',  'quantity', 'public_id')  # adjust fields as per model
-    readonly_fields = ('public_id',)
-    search_fields = ( "public_id", "name", )
-    fields = ('name','quantity', 'description', 'public_id')
 
 
 @admin.register(PasswordResetEvent)
@@ -658,3 +390,7 @@ class AuditLogAdmin(admin.ModelAdmin):
     #     # No department → restrict 
 
     # -----------------------------
+
+# -----------------------------------------------------
+# Agreement Inlines
+# -----------------------------------------------------
